LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/element-request-fullscreen-and-exit-iframe-expected.txt


FAIL Checks that the fullscreenchange events fire in right order when entering fullscreen assert_equals: expected Element node <iframe allowfullscreen="" src="about:blank"></iframe> but got Element node <body></body>
PASS Checks that the fullscreenchange events fire in right order when exiting fullscreen
